+++ apps/playground-web/src/components/in-app-wallet/sponsored-tx.tsx
@@ -12,52 +12,9 @@ import {
   useReadContract,
 } from "thirdweb/react";
 import { THIRDWEB_CLIENT } from "../../lib/client";
-import { CodeExample } from "../code/code-example";
 import { editionDropContract, editionDropTokenId } from "./constants";
 
-export function SponsoredInAppTx() {
-  return (
-    <>
-      <div className="space-y-2">
-        <h2 className="text-2xl sm:text-3xl font-semibold tracking-tight">
-          Signless Sponsored Transactions
-        </h2>
-        <p className="max-w-[600px]">
-          With in-app wallets, users don&apos;t need to confirm every
-          transaction.
-          <br />
-          Combine it with smart account flag to cover gas costs for the best UX.
-        </p>
-      </div>
-      <CodeExample
-        preview={<SponsoredInAppTxPreview />}
-        code={`import { inAppWallet } from "thirdweb/wallets";
-  import { claimTo } from "thirdweb/extensions/erc1155";
-  import { ConnectButton, TransactionButton } from "thirdweb/react";
-
-  
-  const wallets = [
-    inAppWallet(
-      // turn on gas sponsorship for in-app wallets
-      { smartAccount: { chain, sponsorGas: true }}
-    )
-  ];
-
-  function App(){
-    return (<>
-<ConnectButton client={client} wallets={wallets} />
-
-{/* signless, sponsored transactions */}
-<TransactionButton transaction={() => claimTo({ contract, to: "0x123...", tokenId: 0n, quantity: 1n })}>Mint</TransactionButton>
-</>);
-};`}
-        lang="tsx"
-      />
-    </>
-  );
-}
-
-function SponsoredInAppTxPreview() {
+export function SponsoredInAppTxPreview() {
   const wallet = useActiveWallet();
   const { disconnect } = useDisconnect();
   useEffect(() => {
