+++ apps/playground-web/src/components/sign-in/hooks.tsx
@@ -9,48 +9,9 @@ import {
 import { shortenAddress } from "thirdweb/utils";
 import { createWallet } from "thirdweb/wallets";
 import { THIRDWEB_CLIENT } from "../../lib/client";
-import { CodeExample } from "../code/code-example";
 import { Button } from "../ui/button";
 
-export function Hooks() {
-  return (
-    <>
-      <div className="space-y-2">
-        <h2 className="text-2xl sm:text-3xl font-semibold tracking-tight">
-          Custom UI
-        </h2>
-        <p className="max-w-[600px]">
-          Build your own connect UI using react hooks.
-          <br />
-          Wallet state management is all handled for you.
-        </p>
-      </div>
-
-      <CodeExample
-        preview={<HooksPreview />}
-        code={`// Using your own UI
-        import { useConnect } from "thirdweb/react";
-        import { createWallet } from "thirdweb/wallets";
-      
-        function App(){
-          const { connect } = useConnect();
-      
-          return (<>
-      <button onClick={() => connect(async () => {
-        // 350+ wallets supported
-        const wallet = createWallet("io.metamask");
-        await wallet.connect({ client });
-        return wallet;
-      })}>Connect with Metamask</button>
-      </>);
-      };`}
-        lang="tsx"
-      />
-    </>
-  );
-}
-
-function HooksPreview() {
+export function HooksPreview() {
   const account = useActiveAccount();
   const wallet = useActiveWallet();
   const connectMutation = useConnect();
